Fix carousel, delete redundancies
