+++ src/data/projectsData.tsx
@@ -15,6 +15,13 @@ let projectsData: { name: string, thumbnail: string, tagline: string, deployed?:
             repo: 'https://github.com/cOdeBedient/prove-it'
         },
         {
+            name: "bLOOMbABY",
+            thumbnail: '/assets/project-thumbs/bloom-baby-thumb.png',
+            tagline: "Plant a 3D flower and watch it grow. Will you breed the next superflower?",
+            deployed: 'https://bloombabyfe-toms-projects-be041e30.vercel.app/',
+            repo: 'https://github.com/2311-capstone-bloom-brothers/bloom-brothers-fe'
+        },
+        {
             name: "Look at This Snake",
             thumbnail: '/assets/project-thumbs/look-at-this-snake-thumb.png',
             tagline: "The sight of snakes makes you crawl in your skin? This dog will help to desensitize you.",
@@ -24,21 +31,14 @@ let projectsData: { name: string, thumbnail: string, tagline: string, deployed?:
         {
             name: "Up & A w a y",
             thumbnail: '/assets/project-thumbs/up-and-away-thumb.png',
-            tagline: "Organize and book trips on this vanilla JavaScript interface.",
+            tagline: "Organize and book trips on this vanilla JavaScript app.",
             deployed: 'https://codebedient.github.io/Up-and-A-w-a-y/',
             repo: 'https://github.com/cOdeBedient/Up-and-A-w-a-y'
         },
         {
-            name: "bLOOMbABY",
-            thumbnail: '/assets/project-thumbs/bloom-baby-thumb.png',
-            tagline: "Plant a 3D flower and watch it grow. Will you breed the next superflower?",
-            deployed: 'https://bloombabyfe-toms-projects-be041e30.vercel.app/',
-            repo: 'https://github.com/2311-capstone-bloom-brothers/bloom-brothers-fe'
-        },
-        {
             name: "Rancid Tomatillos",
             thumbnail: '/assets/project-thumbs/rancid-tomatillos-thumb.png',
-            tagline: "A React interface for organizing and selecting movies",
+            tagline: "A React interface for organizing and selecting movies.",
             deployed: 'https://rancid-tomatillos-kappa.vercel.app/',
             repo: 'https://github.com/cOdeBedient/rancidTomatillos'
         }
