--- src/version-1-components/Footer.styled.tsx
@@ -1,21 +1,0 @@
-import styled from "styled-components"
-
-export const StyledFooter = styled.section`
-    width: 47vw;
-    padding: 20px;
-    /* background-image: url(/assets/bright-flower-background.jpg); */
-    display: flex;
-    justify-content: space-between;
-    margin-top: 50px;
-
-    button {
-        width: 200px;
-        border: none;
-        margin-left: 5px;
-        margin-right: 5px;
-        border-radius: 5px;
-        font-size: 15px;
-        font-weight: bold;
-        background-color: #d690a8;
-    }
-`
