--- src/version-1-components/Footer.tsx
@@ -1,15 +1,0 @@
-import { StyledFooter } from './Footer.styled'
-
-const Footer: React.FC = () => {
-    return (
-        <>
-            <StyledFooter>
-                    <button>github</button>
-                    <button>instagram</button>
-                    <button>linkedIn</button>
-            </StyledFooter>
-        </>
-    )
-}
-
-export default Footer
